Adding member functions to assemble matrix coeffcients and boundary conditions
